requirements.txt
numpy>=1.26,<2
fastapi==0.110.0
uvicorn[standard]==0.29.0
qdrant-client==1.9.1
sentence-transformers==2.7.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1